+++ huebridgeemulator/tools/__init__.py
@@ -1,4 +1,5 @@
 import socket
+import smtplib
 
 
 def getIpAddress():
@@ -16,3 +17,21 @@ def generateSensorsState(bridge_config, sensors_state):
                     sensors_state[sensor]["state"].update({key: "2017-01-01T00:00:00"})
 
     return bridge_config, sensors_state
+
+
+def sendEmail(triggered_sensor):
+    TEXT = "Sensor " + triggered_sensor + " was triggered while the alarm is active"
+    # Prepare actual message
+    message = """From: %s\nTo: %s\nSubject: %s\n\n%s
+    """ % (bridge_config["alarm_config"]["mail_from"], ", ".join(bridge_config["alarm_config"]["mail_recipients"]), bridge_config["alarm_config"]["mail_subject"], TEXT)
+    try:
+        server_ssl = smtplib.SMTP_SSL(bridge_config["alarm_config"]["smtp_server"], bridge_config["alarm_config"]["smtp_port"])
+        server_ssl.ehlo() # optional, called by login()
+        server_ssl.login(bridge_config["alarm_config"]["mail_username"], bridge_config["alarm_config"]["mail_password"])
+        server_ssl.sendmail(bridge_config["alarm_config"]["mail_from"], bridge_config["alarm_config"]["mail_recipients"], message)
+        server_ssl.close()
+        print("successfully sent the mail")
+        return True
+    except:
+        print("failed to send mail")
+        return False
